# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: LabFlow API
  description: |
    LabFlow 实验室管理系统 API 文档
    - Base URL: http://localhost:8000
    - 认证方式: JWT Token (HttpOnly Cookie)
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development Server
paths:
...
  /api/v1/labs:
    get:
      summary: 获取实验室列表
      description: 分页获取实验室列表
      tags:
        - Lab
      security:
        - CookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: 页码
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
          description: 每页数量
        - name: status
          in: query
          schema:
            type: integer
            enum: [0, 1, 2]
          description: 实验室状态筛选 (0-正常, 1-维护, 2-停用)
        - name: tag_id
          in: query
          schema:
            type: integer
          description: 标签ID筛选
      responses:
        '200':
          description: 成功
          content:
            application/json:
              example:
                code: 200
                message: success
                data:
                  items:
                    - id: 1
                      name: 化学实验室 A
                      address: A 楼 302
                      capacity: 20
                      status: 0
                      tag_id: 1
                      description: 主要用于化学实验教学和科研
                      created_at: "2024-01-15T10:30:00Z"
                      updated_at: "2024-01-15T10:30:00Z"
                  pagination:
                    page: 1
                    page_size: 20
                    total: 10
                    total_pages: 1
    post:
      summary: 创建实验室
      description: 创建新实验室（管理员专用）
      tags:
        - Lab
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabCreate'
      responses:
        '201':
          description: 创建成功
        '403':
          description: 权限不足
  /api/v1/labs/{lab_id}:
    get:
      summary: 获取实验室详情
      description: 根据ID获取实验室详细信息
      tags:
        - Lab
      security:
        - CookieAuth: []
      parameters:
        - name: lab_id
          in: path
          required: true
          schema:
            type: integer
          description: 实验室ID
      responses:
        '200':
          description: 成功
          content:
            application/json:
              example:
                code: 200
                message: success
                data:
                  id: 1
                  name: 化学实验室 A
                  address: A 楼 302
                  capacity: 20
                  status: 0
                  tag_id: 1
                  description: 主要用于化学实验教学和科研
                  created_at: "2024-01-15T10:30:00Z"
                  updated_at: "2024-01-15T10:30:00Z"
        '404':
          description: 实验室不存在
    put:
      summary: 更新实验室
      description: 更新实验室信息（管理员专用）
      tags:
        - Lab
      security:
        - CookieAuth: []
      parameters:
        - name: lab_id
          in: path
          required: true
          schema:
            type: integer
          description: 实验室ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabUpdate'
      responses:
        '200':
          description: 更新成功
        '404':
          description: 实验室不存在
    delete:
      summary: 删除实验室
      description: 删除实验室（管理员专用，软删除）
      tags:
        - Lab
      security:
        - CookieAuth: []
      parameters:
        - name: lab_id
          in: path
          required: true
          schema:
            type: integer
          description: 实验室ID
      responses:
        '200':
          description: 删除成功
        '404':
          description: 实验室不存在
  /api/v1/labs/{lab_id}/managers:
    get:
      summary: 获取实验室负责人列表
      description: 获取实验室的所有负责人
      tags:
        - Lab
      security:
        - CookieAuth: []
      parameters:
        - name: lab_id
          in: path
          required: true
          schema:
            type: integer
          description: 实验室ID
      responses:
        '200':
          description: 成功
          content:
            application/json:
              example:
                code: 200
                message: success
                data:
                  - id: 1
                    user_id: 5
                    name: 李四
                    phone: "13900139000"
    post:
      summary: 添加实验室负责人
      description: 为实验室添加负责人（管理员专用）
      tags:
        - Lab
      security:
        - CookieAuth: []
      parameters:
        - name: lab_id
          in: path
          required: true
          schema:
            type: integer
          description: 实验室ID
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          description: 用户ID
      responses:
        '200':
          description: 添加成功
        '400':
          description: 该用户已是实验室负责人
        '404':
          description: 实验室或用户不存在
    delete:
      summary: 移除实验室负责人
      description: 移除实验室负责人（管理员专用）
      tags:
        - Lab
      security:
        - CookieAuth: []
      parameters:
        - name: lab_id
          in: path
          required: true
          schema:
            type: integer
          description: 实验室ID
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          description: 用户ID
      responses:
        '200':
          description: 移除成功
        '404':
          description: 该用户不是实验室负责人
...
  /api/v1/reservations/{reservation_id}:
    get:
      summary: 获取预约详情
      description: 获取预约详情（本人或管理员）
      tags:
        - Reservation
      security:
        - CookieAuth: []
      parameters:
        - name: reservation_id
          in: path
          required: true
          schema:
            type: integer
          description: 预约ID
      responses:
        '200':
          description: 成功
          content:
            application/json:
              example:
                code: 200
                message: success
                data:
                  id: 1
                  user_id: 3
                  user_name: 张三
                  lab_id: 1
                  lab_name: 化学实验室A
                  start_time: "2024-01-20T09:00:00Z"
                  end_time: "2024-01-20T12:00:00Z"
                  purpose: 实验教学
                  status: 0
                  current_level: 1
                  created_at: "2024-01-15T10:30:00Z"
                  updated_at: "2024-01-15T10:30:00Z"
        '404':
          description: 预约不存在
        '403':
          description: 无权查看
    put:
      summary: 更新预约
      description: 更新预约内容（仅限审批中的预约，本人可操作）
      tags:
        - Reservation
      security:
        - CookieAuth: []
      parameters:
        - name: reservation_id
          in: path
          required: true
          schema:
            type: integer
          description: 预约ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                start_time:
                  type: string
                  format: date-time
                  description: 开始时间
                end_time:
                  type: string
                  format: date-time
                  description: 结束时间
                purpose:
                  type: string
                  description: 使用目的
      responses:
        '200':
          description: 更新成功
        '400':
          description: 更新失败
        '403':
          description: 无权操作
    delete:
      summary: 取消预约
      description: 取消预约（本人或管理员）
      tags:
        - Reservation
      security:
        - CookieAuth: []
      parameters:
        - name: reservation_id
          in: path
          required: true
          schema:
            type: integer
          description: 预约ID
      responses:
        '200':
          description: 取消成功
        '400':
          description: 当前状态无法取消
...
  /api/v1/approvals/pending:
    get:
      summary: 获取待我审批的预约
      description: 获取当前用户需要审批的预约列表（一级和二级）
      tags:
        - Approval
      security:
        - CookieAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              example:
                code: 200
                message: success
                data:
                  level1_pending:
                    - id: 1
                      user_name: 张三
                      lab_id: 1
                      lab_name: 化学实验室A
                      start_time: "2024-01-20T09:00:00Z"
                      end_time: "2024-01-20T12:00:00Z"
                      purpose: 实验教学
                      created_at: "2024-01-15T10:30:00Z"
                  level2_pending:
                    - id: 2
                      user_name: 李四
                      lab_id: 2
                      lab_name: 物理实验室B
                      start_time: "2024-01-21T14:00:00Z"
                      end_time: "2024-01-21T17:00:00Z"
                      purpose: 科研实验
                      created_at: "2024-01-16T11:00:00Z"
  /api/v1/approvals/reservations/{reservation_id}/approve:
    post:
      summary: 审批通过
      description: 审批通过预约（一级或二级审批人）
      tags:
        - Approval
      security:
        - CookieAuth: []
      parameters:
        - name: reservation_id
          in: path
          required: true
          schema:
            type: integer
          description: 预约ID
        - name: level
          in: query
          required: true
          schema:
            type: integer
            enum: [1, 2]
          description: 审批级别 (1-一级审批, 2-二级审批)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
                  description: 审批意见
      responses:
        '200':
          description: 审批通过
          content:
            application/json:
              example:
                code: 200
                message: 审批通过
                data:
                  id: 1
                  status: 0
                  current_level: 2
        '400':
          description: 审批失败
          content:
            application/json:
              example:
                code: 400
                message: 您不是该实验室的负责人
                data: null
  /api/v1/approvals/reservations/{reservation_id}/reject:
    post:
      summary: 审批拒绝
      description: 拒绝预约（一级或二级审批人）
      tags:
        - Approval
      security:
        - CookieAuth: []
      parameters:
        - name: reservation_id
          in: path
          required: true
          schema:
            type: integer
          description: 预约ID
        - name: level
          in: query
          required: true
          schema:
            type: integer
            enum: [1, 2]
          description: 审批级别 (1-一级审批, 2-二级审批)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
                  description: 拒绝原因
      responses:
        '200':
          description: 已拒绝
          content:
            application/json:
              example:
                code: 200
                message: 已拒绝
                data:
                  id: 1
                  status: 2
                  current_level: 1
        '400':
          description: 操作失败
  /api/v1/approvals/reservations/{reservation_id}:
    get:
      summary: 获取预约的审批记录
      description: 获取指定预约的所有审批记录
      tags:
        - Approval
      security:
        - CookieAuth: []
      parameters:
        - name: reservation_id
          in: path
          required: true
          schema:
            type: integer
          description: 预约ID
      responses:
        '200':
          description: 成功
          content:
            application/json:
              example:
                code: 200
                message: success
                data:
                  - id: 1
                    reservation_id: 1
                    approver_id: 5
                    approver_name: 王五
                    level: 1
                    status: 0
                    comment: 同意使用
                    approved_at: "2024-01-16T09:00:00Z"
                    created_at: "2024-01-16T08:30:00Z"
                  - id: 2
                    reservation_id: 1
                    approver_id: 6
                    approver_name: 赵六
                    level: 2
                    status: 0
                    comment: 实验室可用
                    approved_at: "2024-01-16T10:00:00Z"
                    created_at: "2024-01-16T09:30:00Z"
components:
  schemas:
...
    LabCreate:
      type: object
      required:
        - name
        - address
        - tag_id
      properties:
        name:
          type: string
          maxLength: 100
          description: 实验室名称
        address:
          type: string
          maxLength: 255
          description: 地址
        capacity:
          type: integer
          description: 容纳人数
        tag_id:
          type: integer
          description: 标签ID
        keyword:
          type: string
          maxLength: 255
          description: 关键词
        description:
          type: string
          description: 实验室说明
    LabUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          description: 实验室名称
        address:
          type: string
          maxLength: 255
          description: 地址
        capacity:
          type: integer
          description: 容纳人数
        status:
          type: integer
          enum: [0, 1, 2]
          description: 状态 (0-正常, 1-维护, 2-停用)
        tag_id:
          type: integer
          description: 标签ID
        keyword:
          type: string
          maxLength: 255
          description: 关键词
        description:
          type: string
          description: 实验室说明
...
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
...